+++ models/macros.go
@@ -10,14 +10,14 @@ type Macros struct {
 	ID              int       `json:"id"`
 	Account_Id      uuid.UUID `json:"account_id"`
 	Date_Created    time.Time `json:"date_created"`
-	Calories        float32   `json:"calories"`
-	Protein         float32   `json:"protein"`
-	Carbs           float32   `json:"carbs"`
-	Fats            float32   `json:"fats"`
-	Total_Calories  float32   `json:"total_calories"`
-	Total_Protein   float32   `json:"total_protein"`
-	Total_Carbs     float32   `json:"total_carbs"`
-	Total_Fats      float32   `json:"total_fats"`
+	Calories        int       `json:"calories"`
+	Protein         int       `json:"protein"`
+	Carbs           int       `json:"carbs"`
+	Fats            int       `json:"fats"`
+	Total_Calories  int       `json:"total_calories"`
+	Total_Protein   int       `json:"total_protein"`
+	Total_Carbs     int       `json:"total_carbs"`
+	Total_Fats      int       `json:"total_fats"`
 	Activity_Lvl_Id uuid.UUID `json:"activity_lvl_id"`
 	Diet_Plan_Id    uuid.UUID `json:"diet_plan_id"`
 	// time.Time SHOULD BE IN ISO STRING
